requirements.txt
streamlit
pytesseract
pillow
opencv-python
numpy
deep-translator
streamlit-cropper
